Revised the constraint controller. Add rotation constraint for a joint. Can also specify a local offset ( either position or rotation ) for the constraint. ex : bml char doctor <sbm:constraint sbm:handle="cdoctor" effector="l_wrist" sbm:constraint-type="rot" target="foo" rot-x="180" rot-z="180" rot-y="10"/>
Revised the position control UI in the debug viewer. Now the user can drag  along each axis to move the object along that axis.
Add rotation control UI to change the orientation for a pawn object. (Shift + LMB)

Revised example-based reach controller ( sbm:bodyreach ). Add procedural IK animation when no motion data is given. Add initial bml command to use the controller. 
ex : bml char doctor <sbm:bodyreach sbm:handle="creach" target="foo"/>

+++ trunk/core/smartbody/smartbody-lib/src/sbm/me_ct_limb.cpp
@@ -95,7 +95,7 @@ bool MeCtLimb::buildJointChain(SrArray<const char*>& joint_names, SrArray<MeCtIK
 
 	// initialize MeCtIKScenario
 	ik.joint_info_list.size(joint_chain.size());
-	for (size_t i=0;i<joint_chain.size();i++)
+	for (unsigned int i=0;i<joint_chain.size();i++)
 	{	
 		ik.joint_info_list[i] = MeCtIKScenarioJointInfo();
 		MeCtIKScenarioJointInfo& joint_info = ik.joint_info_list[i];
@@ -127,10 +127,10 @@ void MeCtLimb::updateMotionFrameToJointChain( SkMotion* motion, int frame )
 {
 	const float *fp = motion->posture(frame);
 	SkChannelArray& channels = motion->channels();
-	int inserted;
-	int size = channels.size();
+	unsigned int inserted;
+	unsigned int size = channels.size();
 
-	for (int i=0; i<size; i++ ) {
+	for (unsigned int i=0; i<size; i++ ) {
 		if ( channels[i].joint && hasJoint(channels[i].joint) )
 		{
 			inserted = channels[i].set ( fp );
